Bring trunk from subversion
Current version: 1.4.4

+++ include/admin.php
@@ -33,10 +33,7 @@ class Sublanguage_admin extends Sublanguage_main {
 		add_action('init', array($this, 'update'), 15);
 		
 		add_action('plugins_loaded', array($this, 'find_current_language'), 11);
-		
-		add_action('init', array($this, 'register_postmeta_keys'), 99);
 		add_filter('get_post_metadata', array($this, 'translate_meta_data'), null, 4);
-		
 		add_filter('the_posts', array($this, 'translate_the_posts'), 10, 2);
 		add_filter('the_posts', array($this, 'hard_translate_posts'), 20, 2);
 		add_filter('the_post', array($this, 'hard_translate_post'));
@@ -44,6 +41,7 @@ class Sublanguage_admin extends Sublanguage_main {
 		add_filter('home_url', array($this,'translate_home_url'), 10, 4); // /!\ MAY BE USED BY INTERNAL FUNCTIONS /!\
 		add_filter('page_link', array($this, 'translate_page_link'), 10, 3);					
 		add_filter('post_type_link', array($this, 'translate_custom_post_link'), 10, 3);
+		add_filter('attachment_link', array($this, 'translate_attachment_link'), 10, 2);
 				
 		add_filter('list_cats', array($this, 'translate_term_name'), 10, 2);
 		add_filter('get_the_terms', array($this, 'translate_post_terms'), 10, 3); // filter in get_the_terms()
@@ -51,6 +49,9 @@ class Sublanguage_admin extends Sublanguage_main {
 		add_filter('single_cat_title', array($this, 'filter_single_term_title')); // filter term title
 		add_filter('single_tag_title', array($this, 'filter_single_term_title')); // filter term title
 		add_filter('get_terms', array($this, 'translate_get_terms'), 10, 3); // translate post list of categories		
+		
+		add_filter('terms_to_edit', array($this, 'terms_to_edit'), 10, 2); // -> added in 1.4.4
+		add_action('create_term', array($this, 'cancel_term'), 11, 3); // -> added in 1.4.4
 		
 		// for ajax
 		add_filter('get_term', array($this, 'translate_get_term'), 10, 2);
@@ -69,9 +70,7 @@ class Sublanguage_admin extends Sublanguage_main {
 		add_filter('preview_post_link', array($this, 'translate_preview_post_link'), null , 2);
 		
 		// javascript for ajax
-		add_action('admin_footer-post.php', array($this, 'print_javascript_ajax_hook'));	
-		add_action('admin_footer-edit.php', array($this, 'print_javascript_ajax_hook'));
-		add_action('admin_footer-edit-tags.php', array($this, 'print_javascript_ajax_hook'));
+		add_action('admin_enqueue_scripts', array($this, 'ajax_enqueue_scripts'));
 		
 		// edit post meta data
 		add_filter('update_post_metadata', array($this, 'update_translated_postmeta'), null, 5);
@@ -81,7 +80,8 @@ class Sublanguage_admin extends Sublanguage_main {
 		// term
 		add_action('edit_term', array($this, 'save_term_translation'), null, 3);
 		add_action('delete_term', array($this, 'delete_term_translations'), 10, 4);
-				
+		add_filter('terms_clauses', array($this, 'terms_clauses'), 10, 3); // -> added in 1.4.4
+			
 		// translate walker for pages dropdown
 		add_filter('list_pages', array($this, 'translate_list_pages'), 10 , 2);
 		
@@ -125,10 +125,17 @@ class Sublanguage_admin extends Sublanguage_main {
 			$options['current_first'] = false;
 			$options['taxonomy'] = array('category');
 			$options['cpt'] = array('post', 'page');
-			$options['version'] = '1.2';
+			$options['version'] = $this->version;
 	
 			update_option($this->option_name, $options);
+			
+		}
 		
+		// -> avoid database request when this option is not saved.
+		if (!get_option($this->translation_option_name)) {
+		
+			update_option($this->translation_option_name, array());
+			
 		}
 		
 		$admins = get_role( 'administrator' );
@@ -176,25 +183,21 @@ class Sublanguage_admin extends Sublanguage_main {
 	 */
 	public function update() {
 		
-		if (!$this->options || $this->options['version'] != $this->version) {
+		if (version_compare($this->version, $this->options['version']) > 0) { //  $this->version > $this->options['version']
 		
-			include( plugin_dir_path( __FILE__ ) . 'update.php');
-		
-			$old_options = get_option('sublanguage');
+			// upgrades start here :
 			
-			if ($old_options) {
-			
-				sublanguage_update_1($this);
-			
-			}
-			
-			if ($this->options['version'] != $this->version) {
+			if (version_compare($this->options['version'], "1.4.3") <= 0) { // $this->options['version'] <= "1.4.3"
 				
-				sublanguage_update_2($this);
+				$this->clean_orphan_terms();
 			
 			}
 			
-			$this->options = get_option($this->option_name);
+			// upgrades end
+			
+			$this->options['version'] = $this->version;
+			
+			update_option($this->option_name, $this->options);
 			
 		}	
 	
@@ -208,7 +211,7 @@ class Sublanguage_admin extends Sublanguage_main {
 	 * @from 1.0
 	 */	
 	public function insert_post($data, $postarr) {
-		
+					
 		if (isset($this->options['cpt'], $data['post_type']) && $this->options['cpt'] && in_array($data['post_type'], $this->options['cpt'])) { // -> only for translatable post
 			
 			if ($this->current_language->ID != $this->options['main']) { 
@@ -240,22 +243,19 @@ class Sublanguage_admin extends Sublanguage_main {
 				
 				}
 				
-				// store translated data
-				$this->sublanguage_data[$this->current_language->ID]['post_title'] = $data['post_title'];
-				$this->sublanguage_data[$this->current_language->ID]['post_content'] = $data['post_content'];
-				$this->sublanguage_data[$this->current_language->ID]['post_excerpt'] = $data['post_excerpt'];
-				$this->sublanguage_data[$this->current_language->ID]['post_name'] = $data['post_name'];
-			
-				// and restore original data
-				$data['post_title'] = $post->post_title;
-				$data['post_content'] = $post->post_content;
-				$data['post_excerpt'] = $post->post_excerpt;
-				$data['post_name'] = $post->post_name;
+				foreach ($this->fields as $field) {
+					
+					// store translated data
+					$this->sublanguage_data[$this->current_language->ID][$field] = $data[$field];
+					
+					// and restore original data
+					$data[$field] = $post->$field;
+				}
 				
 			}
 		
 		}
-
+			
 		return $data;
 	
 	}
@@ -270,7 +270,9 @@ class Sublanguage_admin extends Sublanguage_main {
 		
 		if ($this->disable_post_filter) return;
 				
-		if (isset($this->current_language->ID) && isset($this->sublanguage_data[$this->current_language->ID])) {
+		if (isset($this->current_language->ID) 
+			&& ($this->current_language->ID != $this->options['main'])
+			&& isset($this->sublanguage_data[$this->current_language->ID])) {
 			
 			if (in_array(get_post($post_id)->post_type, $this->options['cpt'])
 				&& current_user_can('edit_post', $post_id)) {
@@ -285,6 +287,15 @@ class Sublanguage_admin extends Sublanguage_main {
 				
  					$this->disable_post_filter = true;
 					
+					/**
+					 * Fire before a translation is updated.
+					 * @param int $post_id. Original post id.
+					 * @param array $translation_data. Translation data.
+					 *
+					 * @from 1.3
+					 */
+					do_action('sublanguage_admin_update_translation', $post_id, $translation_data);
+					
 					wp_update_post($translation_data);
 					
  					$this->disable_post_filter = false;
@@ -296,6 +307,15 @@ class Sublanguage_admin extends Sublanguage_main {
 					$translation_data['post_parent'] = $post_id;
 					$translation_data['post_type'] = $this->post_translation_prefix.$this->current_language->ID;
 					$translation_data['post_status'] = 'publish'; // or inherit ?
+					
+					/**
+					 * Fire before a translation is created.
+					 * @param int $post_id. Original post id.
+					 * @param array $translation_data. Translation data.
+					 *
+					 * @from 1.3
+					 */
+					do_action('sublanguage_admin_create_translation', $post_id, $translation_data);
 					
 					$translation_id = wp_insert_post($translation_data);
 					
@@ -321,12 +341,16 @@ class Sublanguage_admin extends Sublanguage_main {
 		
 		foreach ($languages as $lng) {
 			
-			$translation = $this->get_post_translation($post_id, $lng->ID);
+			if ($lng->ID != $this->options['main']) {
 			
-			if ($translation) {
+				$translation = $this->get_post_translation($post_id, $lng->ID);
+			
+				if ($translation) {
 				
-				wp_delete_post($translation->ID, true);
+					wp_delete_post($translation->ID, true);
 			
+				}
+				
 			}
 			
 		}
@@ -371,37 +395,7 @@ class Sublanguage_admin extends Sublanguage_main {
 		
 	}
 
-	/**
-	 * Print javascript for interceptig ajax request
-	 *
-	 * Hook for 'admin_footer-{...}'
-	 *
-	 * @from 1.1
-	 */
-	public function print_javascript_ajax_hook() {
-		
-		if (isset($this->current_language->ID) && $this->current_language->ID != $this->options['main']) {
-		
-			$params = $this->language_query_var.'='.$this->current_language->post_name;
-			$params .= '&language_switch_nonce='.wp_create_nonce('language_switch_action');
-			
-			// -> add nonce ?
-		
-?>
-<script type="text/javascript">
-	//<![CDATA[
-		jQuery(document).ready(function($) {
-			$( document ).ajaxSend(function( event, jqxhr, settings ) {
-				settings.url += ((settings.url.indexOf("?") > -1) ? "&" : "?") + "<?php echo $params; ?>";
-			});
-		});
-	//]]>
-</script>
-<?php
-		
-		}
-		
-	}
+
 
 
 
@@ -515,31 +509,35 @@ class Sublanguage_admin extends Sublanguage_main {
 		if ($this->disable_postmeta_filter) return $null;
 		
 		$post = get_post($object_id);
-
-		if ($this->get_language_by_type($post->post_type)) {
+		
+		if ($post) {
+			
+			if ($this->get_language_by_type($post->post_type)) {
 	
-			return true; // -> exit
+				return true; // -> exit
 			
-		} else if (in_array($post->post_type, $this->options['cpt'])) {
+			} else if (in_array($post->post_type, $this->options['cpt'])) {
 			
-			$translatable = in_array($meta_key, $this->postmeta_keys) || apply_filters('sublanguage_translatable_postmeta', false, $meta_key, $object_id);
+				$translatable = in_array($meta_key, $this->postmeta_keys) || apply_filters('sublanguage_translatable_postmeta', false, $meta_key, $object_id);
 		
-			if ($translatable) {
+				if ($translatable) {
 				
-				if ($this->current_language->ID != $this->options['main']) {
+					if ($this->current_language->ID != $this->options['main']) {
 		
-					$translation = $this->get_post_translation($object_id, $this->current_language->ID);
+						$translation = $this->get_post_translation($object_id, $this->current_language->ID);
 		
-					$this->disable_postmeta_filter = true;
+						$this->disable_postmeta_filter = true;
 		
-					delete_metadata('post', $translation->ID, $meta_key, $meta_value, $delete_all);
+						delete_metadata('post', $translation->ID, $meta_key, $meta_value, $delete_all);
 		
-					$this->disable_postmeta_filter = false;
+						$this->disable_postmeta_filter = false;
 		
-					return true; // -> exit;
+						return true; // -> exit;
 			
+					}
+				
 				}
-				
+			
 			}
 			
 		}
@@ -650,14 +648,11 @@ class Sublanguage_admin extends Sublanguage_main {
 	 */
 	public function delete_term_translations($term_id, $tt_id, $taxonomy, $deleted_term) {
 		
-		$original_term = get_term_by('id', $term_id, $taxonomy);
-		$this->enqueue_terms($original_term->term_id);
-		
 		$languages = $this->get_languages();
 		
 		foreach ($languages as $language) {
 		
-			$translation = $this->get_term_translation($original_term, $taxonomy, $language->ID);
+			$translation = $this->get_term_translation($deleted_term, $deleted_term->taxonomy, $language->ID);
 			
 			if ($translation) {
 				
@@ -666,8 +661,242 @@ class Sublanguage_admin extends Sublanguage_main {
 			}
 			
 		}
-		
 
+	}
+
+	
+	
+	/**	 
+	 * When a post with tags is saved while not in main language, a new term with the translated name will be created instead. This cannot be prevented.
+	 * This fontion delete this duplicated term, then reassign the post to the right term.
+	 * Totaly ugly but no alternative
+	 *
+	 * filter for 'create_term'
+	 *
+	 * @from 1.4.4
+	 */
+	public function cancel_term($term_id, $tt_id, $taxonomy) {
+		
+		if ($this->current_language->ID != $this->options['main']) { // -> not main language
+		
+			if (in_array($taxonomy, $this->options['taxonomy'])) {  // -> translatable taxonomy
+				
+				$term = get_term($term_id, $taxonomy);
+				
+				$translation_terms = get_terms($this->term_translation_prefix.$this->current_language->ID, array(
+					'name' => $term->name,
+					'hide_empty' => false
+				));
+
+				foreach ($translation_terms as $translation_term) {
+					
+					$original_term = get_term($translation_term->parent, $taxonomy);
+
+						
+					if (!empty($original_term)) { // -> ok, there is already a translation for this term
+					
+						wp_delete_term($term->term_id, $taxonomy);
+						
+						// -> now reassign the post...
+						
+						// ... for some reason it already good!
+						
+					}
+					
+					break;
+					
+				}
+				
+			}
+		
+		}
+		
+	}
+	
+	/**	 
+	 * Translate terms used in tags meta box
+	 *
+	 * filter for 'terms_to_edit'
+	 *
+	 * @from 1.4.4
+	 */
+	public function terms_to_edit($terms_to_edit, $taxonomy) {
+		
+		if ($this->current_language->ID != $this->options['main'] && in_array($taxonomy, $this->options['taxonomy'])) { // -> not main language && translatable taxonomy
+			
+			$terms = get_terms($taxonomy, array(
+				'name' => array_map('trim', explode(',', $terms_to_edit)),
+				'hide_empty' => false
+			));
+			
+			foreach ($terms as $term) {
+				
+				$term_ids[] = $term->term_id;
+				
+			}
+			
+			$this->enqueue_terms($term_ids, $this->current_language->ID);
+			
+			$translations = array();
+			
+			foreach ($terms as $term) {
+				
+				$translations[] = $this->translate_term_field($term, $taxonomy, $this->current_language->ID, 'name', $term->name);
+			
+			}
+			
+			$terms_to_edit = esc_attr( join( ',', $translations ) );
+			
+		}
+		
+		return $terms_to_edit;
+		
+	}
+	
+	/**	 
+	 * when terms are queried by name, join translations to the query.
+	 * This is needed in admin when receiving a translated post tags 
+	 *
+	 * filter for 'terms_clauses'
+	 *
+	 * @from 1.4.4
+	 */
+	public function terms_clauses($pieces, $taxonomies, $args) {
+		global $wpdb;
+				
+		if ($this->current_language->ID != $this->options['main'] 		// -> not main language
+			&& !array_diff($taxonomies, $this->options['taxonomy'])) {	// -> translatable taxonomy
+			
+			$tr_args = array();
+			
+			if (isset($args['name']) && $args['name']) {
+				
+				$tr_args = array(
+					'hide_empty' => false,
+					'name' => $args['name']
+				);
+				
+			} else if (isset($args['name__like']) && $args['name__like']) {
+				
+				$tr_args = array(
+					'hide_empty' => false,
+					'name__like' => $args['name__like']
+				);
+				
+			}
+			
+			if ($tr_args) {
+			
+				$translation_terms = get_terms($this->term_translation_prefix.$this->current_language->ID, $tr_args);
+			
+				if (!$translation_terms) { // -> find in other languages
+				
+					$languages = $this->get_languages();
+					$translation_post_types = array();
+				
+					foreach ($languages as $lng) {
+					
+						if ($lng->ID != $this->options['main'] && $lng->ID != $this->current_language->ID) {
+					
+							$translation_post_types[] = esc_sql($this->post_translation_prefix.$lng->ID);
+						
+						}
+				
+					}
+				
+					if ($translation_post_types) {
+				
+						$translation_terms = get_terms($translation_post_types, $tr_args);
+				
+					}
+				
+				}
+			
+				$translation_parents = array();
+				
+				foreach ($translation_terms as $translation_term) {
+				
+					$translation_parents[] = $translation_term->parent;
+				
+				}
+			
+				if ($translation_parents) {
+				
+					$pieces['where'] = "(" . $pieces['where'] . ") OR (t.term_id in (" . implode(', ', array_map('intval', $translation_parents)) . "))";
+					
+				}
+			
+			}
+
+		}
+		
+		return $pieces;
+		
+	}
+	
+	/**	 
+	 * Clean orphan terms
+	 *
+	 * @from 1.4.4
+	 */
+	public function clean_orphan_terms() {
+		global $wpdb;
+		
+		$languages = $this->get_languages();
+		$translation_post_types = array();
+	
+		foreach ($languages as $lng) {
+		
+			if ($lng->ID != $this->options['main']) {
+		
+				$translation_post_types[] = esc_sql($this->post_translation_prefix.$lng->ID);
+			
+			}
+	
+		}
+		
+		$translation_terms = array();
+		
+		if ($translation_post_types) {
+				
+			$translation_terms = get_terms($translation_post_types, array(
+				'hide_empty' => false,
+				'fields' => 'id=>parent'
+			));
+		
+		}
+		
+		if ($translation_terms) {
+		
+			$original_terms = get_terms($this->options['taxonomy'], array(
+				'hide_empty' => false,
+				'include' => array_values($translation_terms),
+				'fields' => 'ids'
+			));
+			
+			$orphans = array_diff($translation_terms, $original_terms);
+			
+			if ($orphans) {
+				
+				// but now we need the taxonomy to delete them...
+				$orphan_terms = get_terms($translation_post_types, array(
+					'hide_empty' => false,
+					'include' => array_keys($orphans)
+				));
+				
+				foreach ($orphan_terms as $orphan_term) {
+					
+					
+					
+					wp_delete_term($orphan_term->term_id, $orphan_term->taxonomy);
+				
+				}
+				
+			}
+			
+			
+		}
+		
 	}
 
 }
